requirements/common.txt
pygame>=1.9
PyYAML>=3.12
python-box>=3.1
